+++ railties/test/generators/integration_test_generator_test.rb
@@ -22,4 +22,15 @@ class IntegrationTestGeneratorTest < Rails::Generators::TestCase
     assert_no_file "test/integration/integration_test_test.rb"
     assert_file "test/integration/integration_test.rb"
   end
+
+  def test_rails_test_unit_railtie_is_undefined
+    original_const = Rails.send(:remove_const, :TestUnitRailtie)
+
+    assert_raises(RuntimeError) do
+      run_generator %w(integration)
+    end
+    assert_no_file "test/integration/integration_test.rb"
+  ensure
+    Rails.const_set(:TestUnitRailtie, original_const)
+  end
 end